First Draft of Pieceweaver
I made the surrounding landscape, trees, and added in some trees. I also attempted to make a teepee, corn crops, and a bucket for gathering. I also tried to make the museum part, so when the player walks to the painting they fall down to the level. (I didn't know how to make them teleport.)